+++ config/settings/local.py
@@ -86,4 +86,5 @@ TEST_RUNNER = 'django.test.runner.DiscoverRunner'
 # Plugin settings
 ENABLED_BACKEND_PLUGINS = [
     'timeline_backend',
+    'example_backend_app',
 ]