+++ config/settings/test.py
@@ -71,6 +71,7 @@ LOGGING['handlers']['console']['level'] = 'ERROR'
 # Plugin settings
 ENABLED_BACKEND_PLUGINS = [
     'timeline_backend',
+    'example_backend_app',
 ]
 
 # Projectroles app settings